// src/app/model/global.ts
export class Global {
    regulatoryEventId: number;
    contractType: number;
    domainContextDictionaryId: string;
    domainContextName: string;
    domainContextTypeId: number;
    domainContextSubTypeId: number;
    domainContextPossibleNameDefinitions: string;
    domainContextPossibleValueDefinitions: string;
    phraseRule: string;
    entityRule: string;
    referenceExamples: string;
}